+++ encoding/base64url.ts
@@ -4,7 +4,7 @@ import * as base64 from "./base64.ts";
 
 /*
  * Some variants allow or require omitting the padding '=' signs:
- * https://en.wikipedia.org/wiki/Base64#URL_applications
+ * https://en.wikipedia.org/wiki/Base64#The_URL_applications
  * @param base64url
  */
 export function addPaddingToBase64url(base64url: string): string {
@@ -17,6 +17,10 @@ export function addPaddingToBase64url(base64url: string): string {
 }
 
 function convertBase64urlToBase64(b64url: string): string {
+  if (!/^[-_A-Z0-9]*?={0,2}$/i.test(b64url)) {
+    // Contains characters not part of base64url spec.
+    throw new TypeError("Failed to decode base64url: invalid character");
+  }
   return addPaddingToBase64url(b64url).replace(/\-/g, "+").replace(/_/g, "/");
 }