+++ node/buffer_test.ts
@@ -103,14 +103,14 @@ Deno.test({
   name: "alloc filled correctly with integer",
   fn() {
     const buffer: Buffer = Buffer.alloc(3, 5);
-    assertEquals(buffer, new Uint8Array([5, 5, 5]));
+    assertEquals(buffer, new Buffer([5, 5, 5]));
   },
 });
 
 Deno.test({
   name: "alloc filled correctly with single character",
   fn() {
-    assertEquals(Buffer.alloc(5, "a"), new Uint8Array([97, 97, 97, 97, 97]));
+    assertEquals(Buffer.alloc(5, "a"), new Buffer([97, 97, 97, 97, 97]));
   },
 });
 
@@ -119,7 +119,7 @@ Deno.test({
   fn() {
     assertEquals(
       Buffer.alloc(11, "aGVsbG8gd29ybGQ=", "base64"),
-      new Uint8Array([104, 101, 108, 108, 111, 32, 119, 111, 114, 108, 100]),
+      new Buffer([104, 101, 108, 108, 111, 32, 119, 111, 114, 108, 100]),
     );
   },
 });
@@ -129,7 +129,7 @@ Deno.test({
   fn() {
     assertEquals(
       Buffer.alloc(4, "64656e6f", "hex"),
-      new Uint8Array([100, 101, 110, 111]),
+      new Buffer([100, 101, 110, 111]),
     );
   },
 });
@@ -149,11 +149,11 @@ Deno.test({
   fn() {
     assertEquals(
       Buffer.alloc(7, new Uint8Array([100, 101])),
-      new Uint8Array([100, 101, 100, 101, 100, 101, 100]),
+      new Buffer([100, 101, 100, 101, 100, 101, 100]),
     );
     assertEquals(
       Buffer.alloc(6, new Uint8Array([100, 101])),
-      new Uint8Array([100, 101, 100, 101, 100, 101]),
+      new Buffer([100, 101, 100, 101, 100, 101]),
     );
   },
 });
@@ -163,7 +163,7 @@ Deno.test({
   fn() {
     assertEquals(
       Buffer.alloc(1, new Uint8Array([100, 101])),
-      new Uint8Array([100]),
+      new Buffer([100]),
     );
   },
 });
@@ -173,11 +173,11 @@ Deno.test({
   fn() {
     assertEquals(
       Buffer.alloc(6, new Buffer([100, 101])),
-      new Uint8Array([100, 101, 100, 101, 100, 101]),
+      new Buffer([100, 101, 100, 101, 100, 101]),
     );
     assertEquals(
       Buffer.alloc(7, new Buffer([100, 101])),
-      new Uint8Array([100, 101, 100, 101, 100, 101, 100]),
+      new Buffer([100, 101, 100, 101, 100, 101, 100]),
     );
   },
 });
@@ -313,7 +313,7 @@ Deno.test({
     buffer2.copy(buffer1);
 
     assertEquals(
-      data1,
+      new Buffer(data1),
       buffer1,
     );
   },