+++ node/fs/_fs_readFile_test.ts
@@ -1,10 +1,10 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-import { readFile } from "./_fs_readFile.ts";
-import * as path from "../../../path/mod.ts";
-import { assert, assertEquals } from "../../../testing/asserts.ts";
+import { readFile } from "./promises.ts";
+import * as path from "../../path/mod.ts";
+import { assert, assertEquals } from "../../testing/asserts.ts";
 
 const moduleDir = path.dirname(path.fromFileUrl(import.meta.url));
-const testData = path.resolve(moduleDir, "..", "testdata", "hello.txt");
+const testData = path.resolve(moduleDir, "..", "_fs", "testdata", "hello.txt");
 
 Deno.test("readFileSuccess", async function () {
   const data: Uint8Array = await readFile(testData);