+++ node/perf_hooks_test.ts
@@ -4,7 +4,9 @@ import { assertEquals } from "../testing/asserts.ts";
 Deno.test({
   name: "[perf_hooks] performance",
   fn() {
-    assertEquals(perfHooks.performance, performance);
+    assertEquals(perfHooks.performance.clearMarks, performance.clearMarks);
+    assertEquals(perfHooks.performance.mark, performance.mark);
+    assertEquals(perfHooks.performance.now, performance.now);
   },
 });